GOT IT TO WORK!!!!!!

+++ src/main/groovy/com/saltstack/jenkins/github/webhooks/Cleaner.groovy
@@ -8,6 +8,7 @@ import hudson.triggers.Trigger;
 import hudson.util.TimeUnit2;
 import org.kohsuke.github.GHException;
 import org.kohsuke.github.GHHook;
+import org.kohsuke.github.GHEvent;
 import org.kohsuke.github.GHRepository;
 import com.cloudbees.jenkins.GitHubRepositoryName;
 import com.cloudbees.jenkins.GitHubRepositoryNameContributor;
@@ -68,7 +69,9 @@ public class Cleaner extends PeriodicWork {
         for (GitHubRepositoryName r : names) {
             for (GHRepository repo : r.resolve()) {
                 try {
-                    removeHook(repo, Trigger.all().get(BranchesTrigger.DescriptorImpl.class).getHookUrl());
+                    removeHook(repo,
+                               Trigger.all().get(BranchesTrigger.DescriptorImpl.class).getHookUrl(),
+                               [GHEvent.CREATE, GHEvent.DELETE]);
                     LOGGER.fine("Removed a hook from "+r+"");
                     continue OUTER;
                 } catch (Throwable e) {
@@ -80,7 +83,7 @@ public class Cleaner extends PeriodicWork {
 
     //Maybe we should create a remove hook method in the Github API
     //something like public void removeHook(String name, Map<String,String> config)
-    private void removeHook(GHRepository repo, URL url) {
+    private void removeHook(GHRepository repo, URL url, List<GHEvent> events) {
         try {
             String urlExternalForm = url.toExternalForm();
             for (GHHook h : repo.getHooks()) {
@@ -88,7 +91,9 @@ public class Cleaner extends PeriodicWork {
                     continue
                 }
                 if (h.getConfig().get("url").equals(urlExternalForm)) {
-                    h.delete();
+                    if ( repo.getEvents() as Set == events as Set ) {
+                        h.delete();
+                    }
                 }
             }
         } catch (IOException e) {
